-Internal: Linux compile fix

+++ playerbot/RandomPlayerbotMgr.cpp
@@ -533,7 +533,7 @@ void RandomPlayerbotMgr::UpdateAIInternal(uint32 elapsed, bool minimal)
 
     if (sPlayerbotAIConfig.hasLog("activity_pid.csv"))
     {
-        SIZE_T virtualMemUsedByMe = 0;
+        double virtualMemUsedByMe = 0;
 #if PLATFORM == PLATFORM_WINDOWS
         PROCESS_MEMORY_COUNTERS_EX pmc;
         GetProcessMemoryInfo(GetCurrentProcess(), (PROCESS_MEMORY_COUNTERS*)&pmc, sizeof(pmc));
